refactorisation instanciation d'objet de parametre
